Creation of /MultivariateAnalyses directory
Reorganisation and consolidation of the code, data, and SI for the PCA, controlled diet PCA, DFA, and climate (environmental) PCA; now all of this can be found in the directory /MultivariateAnalyses. Old and redundant files have been removed (archived). Links in the big .Rmd have been redirected, and the markdown has been knitted.